--- sktime/classification/interval_based/_vector/__init__.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-__all__ = [
-    "ContinuousIntervalTree",
-]
-
-from sktime.classification.interval_based._vector._continuous_interval_tree import (
-    ContinuousIntervalTree,
-)
