+++ sktime/contrib/_continuous_interval_tree.py
